Added EnvironmentModule, deleted the package XR Interaction ToolKit Version 3.0.8, Added the package XR Interaction ToolKit Version 2.6.3, Added new VR background environment(Coffee Shop), Fixed the error from Assets/Scripts/PlayBack/SubtitleManager.cs
in Assets/Scripts/PlayBack/SubtitleManager.cs line 16:
Cast GetCurrentTime() to int to fix type mismatch in getLine() call. Play Mode runs successfully now.

+++ Lecture_Translator_Frontend/Assets/Scripts/Playback/SubtitleManager.cs
@@ -13,7 +13,7 @@ public class SubtitleManager : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        string currentLine = transcriptManager.GetTranscript(currentLanguage).getLine(playbackManager.GetCurrentTime());
+        string currentLine = transcriptManager.GetTranscript(currentLanguage).getLine((int)playbackManager.GetCurrentTime());
         //lecturePlayerWindow
     }
